f64 is a little too big...

+++ src/protocol/field.rs
@@ -76,7 +76,7 @@ pub struct Field {
 
 impl Field {
 
-    fn get_unit(&self) -> f64 {
+    fn get_unit(&self) -> f32 {
         match self.unit_value.as_str() {
             "0.001" => 0.001,
             "0.01" => 0.01,
@@ -108,7 +108,7 @@ impl Field {
                     },
                     _ => (input, 0),
                 };
-                let value = value as f64 * self.get_unit();
+                let value = value as f32 * self.get_unit();
                 (input, Some(FieldType::Int(value)))
             },
             "bitRange" => {
@@ -134,7 +134,7 @@ impl Field {
             },
             "long" => {
                 let (input, value) = be_i32(input)?;
-                let value = value as f64 * self.get_unit();
+                let value = value as f32 * self.get_unit();
                 (input, Some(FieldType::Long(value)))
             },
             "onePosition" => {
@@ -152,7 +152,7 @@ impl Field {
             },
             "uInt" => {
                 let (input, value) = be_u16(input)?;
-                let value = value as f64 * self.get_unit();
+                let value = value as f32 * self.get_unit();
                 (input, Some(FieldType::UInt(value)))
             },
             _ => (input, None),
